Administration

+++ app/Http/Controllers/PostsController.php
@@ -6,22 +6,31 @@ use App\Notifications\PostCreated;
 use App\Notifications\PostDeleted;
 use App\Notifications\PostEdited;
 use App\Post;
-use App\Post_tag;
+use App\PostTag;
 use App\Tag;
-use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 
 class PostsController extends Controller
 {
     public function __construct()
     {
         $this->middleware('auth');
-        $this->middleware('can:update,post')->except(['index', 'userPosts', 'adminIndex', 'create', 'store']);
+    }
+
+    public function validateRequest($request, $post)
+    {
+        return $request->validate([
+            'code' => ['required', 'regex:/[a-zA-Z0-9_-]+/', Rule::unique('posts')->ignore($post->id)],
+            'name' => 'required|min:5|max:100',
+            'description' => 'required|max:255',
+            'text' => 'required'
+        ]);
     }
 
     public function index()
     {
-        $posts = Post::with('tags')->latest()->get();
+        $posts = Post::with('tags')->where('published', 1)->latest()->get();
         return view('/index', compact('posts'));
     }
 
@@ -31,11 +40,6 @@ class PostsController extends Controller
         return view('/posts.index', compact('posts'));
     }
 
-    public function adminIndex() {
-        $posts = Post::with('tags')->latest()->get();
-        return view('/posts.admin-index', compact('posts'));
-    }
-
     public function create()
     {
         return view('/posts.create');
@@ -43,12 +47,7 @@ class PostsController extends Controller
 
     public function store(Request $request)
     {
-        $attr = $request->validate([
-            'code' => 'required|unique:posts|regex:/[a-zA-Z0-9_-]+/',
-            'name' => 'required|min:5|max:100',
-            'description' => 'required|max:255',
-            'text' => 'required'
-        ]);
+        $attr = $this->validateRequest($request, new Post());
 
         if ($request->has('published')) {
             $attr['published'] = 1;
@@ -58,29 +57,42 @@ class PostsController extends Controller
 
         $post = Post::create($attr);
 
+        if (!is_null($request['tags'])) {
+            $requestTags = explode(', ', $request['tags']);
+            foreach ($requestTags as $tag) {
+                $tag = Tag::firstOrCreate(['name' => $tag]);
+                $post->tags()->attach($tag);
+            }
+        }
+
         sendMailNotifyToAdmin(new PostCreated($post));
         flash( 'Post created successfully');
+        pushNotification('Post created successfully');
 
         return redirect('/');
     }
 
     public function show(Post $post)
     {
+        $this->authorize('view', $post);
+
         return view('/posts.show', compact('post'));
     }
 
     public function edit(Post $post)
     {
+        $this->authorize('update', $post);
+
         return view('/posts.edit', compact('post'));
     }
 
     public function update(Request $request, Post $post)
     {
-        $values = $request->validate([
-            'name' => 'required|min:5|max:100',
-            'description' => 'required|max:255',
-            'text' => 'required'
-        ]);
+        $this->authorize('update', $post);
+
+        $values = $this->validateRequest($request, $post);
+
+        $request->has('published') ? $values['published'] = 1 : $values['published'] = 0;
 
         $post->update($values);
 
@@ -105,24 +117,28 @@ class PostsController extends Controller
         if ($deleteTags->isNotEmpty()) {
             foreach ($deleteTags as $tag) {
                 $post->tags()->detach($tag);
-                $isLastTag = Post_tag::where('tag_id', $tag->id)->first();
+                $isLastTag = PostTag::where('tag_id', $tag->id)->first();
                 if (!$isLastTag) $tag->delete();
             };
         }
 
         sendMailNotifyToAdmin(new PostEdited($post));
         flash( 'Post edited successfully');
+        pushNotification('Post edited successfully');
 
         return back();
     }
 
     public function destroy(Post $post)
     {
+        $this->authorize('delete', $post);
+
         $post->delete();
 
         sendMailNotifyToAdmin(new PostDeleted($post));
-        flash( 'Post delete successfully');
+        flash( 'Post deleted successfully');
+        pushNotification('Post deleted successfully');
 
-        return redirect('/posts');
+        return back();
     }
 }
